+++ docs/source/release_notes.rst
@@ -8,6 +8,8 @@ Release Notes
         * Deleted the ``return_pandas`` flag from our demo data loaders :pr:`2181`
         * Separated `custom_hyperparameters` from pipelines and added them as an argument to `AutoMLSearch :pr:`2317`
     * Documentation Changes
+    * Testing Changes
+        * Ignoring ``test_saving_png_file`` while building conda package :pr:`2323`
 
 .. warning::